refactor: share the public-token JWT scope decoder; make @trigger.dev/plugins internal
Move buildJwtAbility into @trigger.dev/plugins as the single decoder for
public-token scope strings, and re-export it from @trigger.dev/rbac so the
built-in fallback and any auth plugin decode tokens identically. Scope ids
split on only the first two colons, so a resource id that itself contains
colons (e.g. a tag like "user:123") is matched in full rather than
truncated at the first segment.

Make @trigger.dev/plugins private (unpublished) and add a source export
condition, so consumers bundle it from source per-commit like
@trigger.dev/core instead of resolving a published version.
