+++ packages/plugins/src/index.ts
@@ -17,7 +17,9 @@ export type {
   AuthenticatedEnvironment,
 } from "./rbac.js";
 
-// Convenience re-exports — gives plugin authors (and the cloud workspace
-// link) one import surface without reaching into @trigger.dev/core
-// directly. Both helpers live in core; this is purely a forwarder.
+export { buildJwtAbility } from "./rbac.js";
+
+// Convenience re-exports — give plugin authors one import surface
+// without reaching into @trigger.dev/core directly. Both helpers live in
+// core; this is purely a forwarder.
 export { sanitizeBranchName, isValidGitBranchName } from "@trigger.dev/core/v3/utils/gitBranch";
